+++ py-rattler/rattler/match_spec/match_spec.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Optional
 
 from rattler.channel.channel import Channel
-from rattler.package.package_name import PackageName
+from rattler.package.package_name_matcher import PackageNameMatcher
 from rattler.rattler import PyMatchSpec
 
 if TYPE_CHECKING:
@@ -40,6 +40,7 @@ class MatchSpec:
 
     1. `name` (i.e. "package name") is required, but its value can be '*'. Its
     position is always outside the key-value brackets.
+    It can also be a glob pattern or a regex if `exact_names_only` is `False`.
     2. If `version` is an exact version, it goes outside the key-value brackets and
     is prepended by `==`. If `version` is a "fuzzy" value (e.g. `1.11.*`), it goes
     outside the key-value brackets with the `.*` left off and is prepended by `=`.
@@ -84,6 +85,7 @@ class MatchSpec:
         self,
         spec: str,
         strict: bool = False,
+        exact_names_only: bool = True,
         experimental_extras: bool = False,
         experimental_conditionals: bool = False,
     ) -> None:
@@ -101,24 +103,31 @@ class MatchSpec:
         MatchSpec("pip >=24.0")
         >>> MatchSpec("pip 24")
         MatchSpec("pip ==24")
-        >>> MatchSpec('python[license=MIT]')
+        >>> MatchSpec("python[license=MIT]")
         MatchSpec("python[license="MIT"]")
+        >>> MatchSpec("foo*", strict=True, exact_names_only=False)
+        MatchSpec("foo*")
+        >>> MatchSpec("^foo.*$", strict=True, exact_names_only=True) # doctest: +IGNORE_EXCEPTION_DETAIL
+        Traceback (most recent call last):
+        InvalidMatchSpecException: only exact package name matchers are allowed. Got ^foo.*$
         >>>
         ```
         """
         if isinstance(spec, str):
-            self._match_spec = PyMatchSpec(spec, strict, experimental_extras, experimental_conditionals)
+            self._match_spec = PyMatchSpec(
+                spec, strict, exact_names_only, experimental_extras, experimental_conditionals
+            )
         else:
             raise TypeError(
                 f"MatchSpec constructor received unsupported type {type(spec).__name__!r} for the 'spec' parameter"
             )
 
     @property
-    def name(self) -> Optional[PackageName]:
+    def name(self) -> Optional[PackageNameMatcher]:
         """
         The name of the package.
         """
-        return PackageName._from_py_package_name(self._match_spec.name)
+        return PackageNameMatcher._from_py_package_name_matcher(self._match_spec.name)
 
     @property
     def version(self) -> Optional[str]:
@@ -227,7 +236,7 @@ class MatchSpec:
         MatchSpec("foo ==3.4")
         >>> MatchSpec.from_nameless(spec, "$foo") # doctest: +IGNORE_EXCEPTION_DETAIL
         Traceback (most recent call last):
-        exceptions.InvalidPackageNameException
+        exceptions.PackageNameMatcherParseException
         >>>
         ```
         """